fix: business logic is now functioning again
