+++ .pylintrc
@@ -91,7 +91,13 @@ disable=raw-checker-failed,
         too-many-locals,
         import-error,
         invalid-name,
-        duplicate-code
+        duplicate-code,
+        wildcard-import,
+        unused-wildcard-import,
+        line-too-long,
+        missing-class-docstring,
+        missing-function-docstring,
+        missing-module-docstring
 
 # Enable the message, report, category or checker with the given id(s). You can
 # either give multiple identifier separated by comma (,) or put this option
