+++ cyclonedx/clients/ion_channel/ion_channel.py
@@ -12,8 +12,7 @@ from botocore.client import BaseClient
 from botocore.exceptions import ClientError
 from requests import Response
 
-from cyclonedx.constants import PYTHON_LOGGING_CONFIG
-from deploy.constants import IC_API_KEY, IC_RULESET_TEAM_ID
+from cyclonedx.constants import IC_API_KEY, IC_RULESET_TEAM_ID, PYTHON_LOGGING_CONFIG
 from cyclonedx.exceptions.ion_channel import IonChannelError
 
 config.fileConfig(PYTHON_LOGGING_CONFIG)
