+++ cyclonedx/dtendpoints.py
@@ -3,7 +3,8 @@ Module to hold constant values for the API
 """
 
 import os
-from deploy.constants import DT_API_BASE, DT_API_PORT
+
+from cyclonedx.constants import DT_API_BASE, DT_API_PORT
 
 
 class DTEndpoints:
