+++ deploy/authorizers/upload_api_key_authorizer.py
@@ -6,7 +6,8 @@ from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_lambda as lambda_
 from constructs import Construct
 
-from deploy.constants import API_KEY_AUTHORIZER_LN, PRIVATE, SBOM_API_PYTHON_RUNTIME
+from cyclonedx.constants import AWS_ACCOUNT_ID, ENVIRONMENT
+from deploy.constants import API_KEY_AUTHORIZER_LN, SBOM_API_PYTHON_RUNTIME
 from deploy.util import DynamoTableManager, create_asset
 
 
@@ -30,11 +31,17 @@ class SBOMUploadAPIKeyAuthorizerLambda(Construct):
             function_name=API_KEY_AUTHORIZER_LN,
             runtime=SBOM_API_PYTHON_RUNTIME,
             vpc=vpc,
-            vpc_subnets=ec2.SubnetSelection(subnet_type=PRIVATE),
+            vpc_subnets=ec2.SubnetSelection(
+                subnet_type=ec2.SubnetType.PRIVATE_WITH_EGRESS,
+            ),
             handler="cyclonedx.handlers.api_key_authorizer_handler",
             code=create_asset(self),
             timeout=Duration.minutes(2),
             memory_size=512,
+            environment={
+                "CDK_DEFAULT_ACCOUNT": AWS_ACCOUNT_ID,
+                "ENVIRONMENT": ENVIRONMENT,
+            },
         )
 
         table_mgr.grant(self.func)
