+++ deploy/enrichment/ion_channel/ion_channel_interface.py
@@ -12,12 +12,11 @@ from aws_cdk import aws_ssm as ssm
 from constructs import Construct
 
 from deploy.constants import (
-    IC_INTERFACE_LN,
-    PRIVATE,
-    SBOM_API_PYTHON_RUNTIME,
     IC_API_BASE,
     IC_API_KEY,
+    IC_INTERFACE_LN,
     IC_RULESET_TEAM_ID,
+    SBOM_API_PYTHON_RUNTIME,
 )
 from deploy.util import create_asset
 
@@ -46,7 +45,9 @@ class IonChannelInterfaceLambda(Construct):
             function_name=IC_INTERFACE_LN,
             runtime=SBOM_API_PYTHON_RUNTIME,
             vpc=vpc,
-            vpc_subnets=ec2.SubnetSelection(subnet_type=PRIVATE),
+            vpc_subnets=ec2.SubnetSelection(
+                subnet_type=ec2.SubnetType.PRIVATE_WITH_EGRESS,
+            ),
             handler="cyclonedx.handlers.ic_interface_handler",
             code=create_asset(self),
             timeout=Duration.minutes(15),
