+++ deploy/user/lambda_objects/user_search.py
@@ -1,22 +1,18 @@
 """
 -> Module to house the User Search Lambda Construct
 """
-from aws_cdk import (
-    aws_cognito as cognito,
-    aws_ec2 as ec2,
-    aws_lambda as lambda_,
-    Duration,
-)
+from aws_cdk import Duration
+from aws_cdk import aws_cognito as cognito
+from aws_cdk import aws_ec2 as ec2
+from aws_cdk import aws_lambda as lambda_
 from aws_cdk.aws_iam import Effect, PolicyStatement
 from constructs import Construct
 
-from cyclonedx.constants import (
+from cyclonedx.constants import AWS_ACCOUNT_ID, ENVIRONMENT
+from deploy.constants import (
+    SBOM_API_PYTHON_RUNTIME,
     USER_POOL_CLIENT_ID_KEY,
     USER_POOL_ID_KEY,
-)
-from deploy.constants import (
-    PRIVATE,
-    SBOM_API_PYTHON_RUNTIME,
     USER_SEARCH_LN,
 )
 from deploy.util import create_asset
@@ -43,7 +39,9 @@ class SBOMUserSearchLambda(Construct):
             function_name=USER_SEARCH_LN,
             runtime=SBOM_API_PYTHON_RUNTIME,
             vpc=vpc,
-            vpc_subnets=ec2.SubnetSelection(subnet_type=PRIVATE),
+            vpc_subnets=ec2.SubnetSelection(
+                subnet_type=ec2.SubnetType.PRIVATE_WITH_EGRESS,
+            ),
             handler="cyclonedx.handlers.user_search_handler",
             code=create_asset(self),
             timeout=Duration.seconds(10),
@@ -51,6 +49,8 @@ class SBOMUserSearchLambda(Construct):
             environment={
                 USER_POOL_ID_KEY: user_pool.user_pool_id,
                 USER_POOL_CLIENT_ID_KEY: user_pool_client.user_pool_client_id,
+                "CDK_DEFAULT_ACCOUNT": AWS_ACCOUNT_ID,
+                "ENVIRONMENT": ENVIRONMENT,
             },
         )
